valeohr/backend/requirements.txt
Django==3.0.1
django-mssql-backend==2.8.1
pyodbc==4.0.30
pytz==2020.5
sqlparse==0.4.1
requests
djangorestframework==3.13.1
markdown
django-filter
djangorestframework-simplejwt==4.7.0    
PyJWT
python-dateutil
django-cors-headers
pyyaml
drf-yasg
django-rest-serializer-field-permissions==2.0
pysftp
django-crontab==0.7.1
APScheduler==3.7.0
